+++ include/iot/uuid.h
@@ -1,20 +1,23 @@
-/**
- * Copyright (c) 2018 rxi
- *
- * This library is free software; you can redistribute it and/or modify it
- * under the terms of the MIT license. See LICENSE for details.
- */
-
-/* Ref: https://github.com/rxi/uuid4 */
+//
+// Copyright (c) 2021 IOTech Ltd
+//
+// SPDX-License-Identifier: Apache-2.0
+//
 #ifndef UUID_H
 #define UUID_H
 
-#define UUID_VERSION "1.0.0"
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 #define UUID_STR_LEN 37
 
 typedef unsigned char uuid_t[16];
 
-void uuid_generate(uuid_t out);
-void uuid_unparse(const uuid_t uu, char *out);
+extern void uuid_generate (uuid_t out);
+extern void uuid_unparse (const uuid_t uuid, char * out);
 
+#ifdef __cplusplus
+}
+#endif
 #endif