+++ RELEASE_NOTES.md
@@ -56,10 +56,14 @@
 
 * `iot_time_usecs`
 
-- Zero length `iot_data_t` arrays now supported
-
 ## Version 1.2.2
 
+- Added support for UUID in data. New functions added:
+
+* `iot_data_alloc_uuid`
+* `iot_data_alloc_uuid_string`
+
+- Zero length `iot_data_t` arrays now supported
 - Data map performance enhancements. Note iteration order is now based on key ordering.
 - Added metadata based JSON object key ordering support. This includes the addition of the function: